feat: component 분리작업
