+++ apps/web/src/apis/train.ts
@@ -1,7 +1,7 @@
 import { AxiosInstance } from 'axios'
 import { ax } from './axios'
 import { StandardResponse } from '@/types/global'
-import { SubwayLineServerModel, TrainMetaData } from '@/types/train'
+import { SubwayLineServerModel, TrainMetaData, TrainRealTimeData } from '@/types/train'
 
 export const getSubwayLines = async () => {
   const { data } = await ax.get<StandardResponse<SubwayLineServerModel>>('/subway-lines')
@@ -14,7 +14,7 @@ export const fetchGetTrains = async (trainId: string) => {
 }
 
 export const fetchGetTrainRealTimeData = async (stationInfo: { stationId?: number; subwayLineId?: number }) => {
-  const { data } = await ax.get<StandardResponse<TrainMetaData>>(
+  const { data } = await ax.get<StandardResponse<TrainRealTimeData>>(
     `/trains/real-times?stationId=${stationInfo?.stationId}&subwayLineId=${stationInfo?.subwayLineId}`
   )
   return data
